modularized database and question creation logic
